Refactor to prepare evaluation
