+++ src/Api/Entity/CraftsmanStatistics.php
@@ -11,7 +11,9 @@
 
 namespace App\Api\Entity;
 
-class CraftsmanStatistics
+use App\Service\Craftsman\Statistics;
+
+class CraftsmanStatistics extends Statistics
 {
     /**
      * @var string
@@ -19,43 +21,14 @@ class CraftsmanStatistics
     private $craftsman;
 
     /**
-     * @var IssueSummary
+     * CraftsmanStatistics constructor.
      */
-    private $issueSummary;
-
-    /**
-     * @var int
-     */
-    public $issueUnreadCount = 0;
-
-    /**
-     * @var int
-     */
-    public $issueOverdueCount = 0;
-
-    /**
-     * @var \DateTime|null
-     */
-    public $nextDeadline;
-
-    /**
-     * @var \DateTime|null
-     */
-    public $lastEmailReceived;
-
-    /**
-     * @var \DateTime|null
-     */
-    public $lastVisitOnline;
-
-    /**
-     * @var \DateTime|null
-     */
-    public $lastIssueResolved;
-
-    public function __construct()
+    public function __construct(string $craftsmanIri, Statistics $statistics)
     {
-        $this->issueSummary = new IssueSummary();
+        parent::__construct();
+
+        $this->craftsman = $craftsmanIri;
+        parent::writeFrom($statistics);
     }
 
     public function getCraftsman(): string
@@ -66,70 +39,5 @@ class CraftsmanStatistics
     public function setCraftsman(string $craftsman): void
     {
         $this->craftsman = $craftsman;
-    }
-
-    public function getIssueSummary(): IssueSummary
-    {
-        return $this->issueSummary;
-    }
-
-    public function getIssueUnreadCount(): int
-    {
-        return $this->issueUnreadCount;
-    }
-
-    public function setIssueUnreadCount(int $issueUnreadCount): void
-    {
-        $this->issueUnreadCount = $issueUnreadCount;
-    }
-
-    public function getIssueOverdueCount(): int
-    {
-        return $this->issueOverdueCount;
-    }
-
-    public function setIssueOverdueCount(int $issueOverdueCount): void
-    {
-        $this->issueOverdueCount = $issueOverdueCount;
-    }
-
-    public function getNextDeadline(): ?\DateTime
-    {
-        return $this->nextDeadline;
-    }
-
-    public function setNextDeadline(?\DateTime $nextDeadline): void
-    {
-        $this->nextDeadline = $nextDeadline;
-    }
-
-    public function getLastEmailReceived(): ?\DateTime
-    {
-        return $this->lastEmailReceived;
-    }
-
-    public function setLastEmailReceived(?\DateTime $lastEmailReceived): void
-    {
-        $this->lastEmailReceived = $lastEmailReceived;
-    }
-
-    public function getLastVisitOnline(): ?\DateTime
-    {
-        return $this->lastVisitOnline;
-    }
-
-    public function setLastVisitOnline(?\DateTime $lastVisitOnline): void
-    {
-        $this->lastVisitOnline = $lastVisitOnline;
-    }
-
-    public function getLastIssueResolved(): ?\DateTime
-    {
-        return $this->lastIssueResolved;
-    }
-
-    public function setLastIssueResolved(?\DateTime $lastIssueResolved): void
-    {
-        $this->lastIssueResolved = $lastIssueResolved;
     }
 }
